+++ index.html
@@ -51,6 +51,7 @@
 	<li><a href="./features/element-hiding/">Element Hiding</a></li>
 	<li><a href="./features/auto-refresh.html">Auto Refresh</a></li>
 	<li><a href="./features/favicon/">Favicon</a></li>
+	<li><a href="./network-error/">Network Errors</a></li>
 </ul>
 
 <h2>Security</h2>
@@ -121,8 +122,6 @@
 	<li><a href='./windows-browser/script-injection/index.html'>Windows Browser: MainWorld Script injection tests</a></li>
 	<li><a href='./windows-browser/script-injection/secureWorld_index.html'>Windows Browser: SecureWorld Script injection tests</a></li>
 	<li><a href='./tools/reach-calculator.html'>Site Breakage Reach Calculator</a></li>
-	<li><a href='./error?code=401'>HTTP 401 error (and others)</a></li>
-	<li><a href='./error/drop'>Simulate connection drop</a></li>
 </ul>
 
 </body>
