+++ server.js
@@ -261,30 +261,6 @@ app.get('/security/address-bar-spoofing/download-redirect', (req, res) => {
     res.redirect(301, 'https://staticcdn.duckduckgo.com/macos-desktop-browser/duckduckgo.dmg');
 });
 
-// Simulate HTTP errors
-app.get('/error', (req, res) => {
-    const code = parseInt(req.query.code);
-
-    if (code === 401) {
-        // This sets a WebView error on Windows
-        res.set('WWW-Authenticate', 'Basic realm="Restricted Area"');
-    }
-
-    if (isNaN(code) || code < 100 || code > 599) {
-        res.statusCode = 200;
-        res.send('Invalid HTTP error code provided. Status not set.');
-    } else {
-        res.statusCode = code;
-        res.send(`HTTP error code set to '${code}'`);
-    }
-    res.end();
-});
-
-// Simulate connection drop
-app.get('/error/drop', (_, res) => {
-    res.destroy();
-});
-
 app.use('/content-scope-scripts/', express.static('node_modules/@duckduckgo/content-scope-scripts/integration-test/test-pages/'));
 
 const blockingRoutes = require('./privacy-protections/request-blocking/server/routes');
@@ -310,3 +286,6 @@ app.use('/viewport', viewportRoutes);
 
 const addressBarSpoofingRoutes = require('./security/address-bar-spoofing/server/routes.js');
 app.use('/security/address-bar-spoofing-download-redirect', addressBarSpoofingRoutes);
+
+const networkErrorRoutes = require('./network-error/server/routes.js');
+app.use('/network-error', networkErrorRoutes);
